+++ src/movingaves_max.cpp
@@ -2,15 +2,15 @@
 using namespace Rcpp;
 
 // [[Rcpp::export]]
-float movingaves_max(NumericVector x, int window) {
+double movingaves_max(NumericVector x, int window) {
   int n = x.size();
-  float sum = 0;
+  double sum = 0;
   NumericVector current(window);
   for (int a = 0; a < window; ++a) {
     current[a] = x[a];
     sum += x[a];
   }
-  float max=sum;
+  double max=sum;
   for (int b = window; b < n; ++b) {
     sum = sum + x[b] - x[b-window];
     if (sum>max) max=sum;
